Aggiunto controllo aggiuntivo quando viene spostata un'immagine. Fix tipo eccezione errata in Checks. Fix redirect in varie jsp. Fix RGX delle descrizioni. Aggiunto metodo doRemoveDiscount in ProdottoDAO. Aggiunto controllo immagine mancante anche in Product.jsp e Shop.jsp. Fix in RecensioneDAO del metodo doExistsByUsername che diventa doExistsByUsername_Prod. Fix del parse in ReviewServlet. Vari cambiamenti estetici. Implementata la modifica dello sconto per un prodotto.

+++ src/main/java/com/tsw/studentsupps/Controller/Admin/AdminDeleteProductServlet.java
@@ -33,13 +33,15 @@ public class AdminDeleteProductServlet extends HttpServlet {
 
         String delProdImageFolder= (String) getServletContext().getAttribute("delProdImageFolder");
         File movedImage= new File(delProdImageFolder, imageToDelete);
-        for(int i= 0; i<1000; i++) {
-            if(!Files.exists(movedImage.toPath())) {
-                Files.move(imageToMove.toPath(), movedImage.toPath());
-                break;
-            }
-            else {
-                movedImage= new File(delProdImageFolder, prodName + i + imageToDelete.substring(imageToDelete.lastIndexOf(".")));
+        if(Files.exists(imageToMove.toPath())) {
+            for(int i= 0; i<1000; i++) {
+                if(!Files.exists(movedImage.toPath())) {
+                    Files.move(imageToMove.toPath(), movedImage.toPath());
+                    break;
+                }
+                else {
+                    movedImage= new File(delProdImageFolder, prodName + i + imageToDelete.substring(imageToDelete.lastIndexOf(".")));
+                }
             }
         }
